--- hooks/src/lib/common/index.ts
@@ -1,1 +1,0 @@
-export { default as useInput } from "./useInput";
