+++ hooks/src/lib/useCardHolder/useCardHolder.ts
@@ -1,7 +1,7 @@
 import { ChangeEvent, useState, KeyboardEvent, FocusEvent } from "react";
-import { useInput } from "../common";
 import Validator from "../utils/validator";
 import { ERROR_MESSAGE } from "../constants";
+import useInput from "../common/useInput";
 
 const useCardHolder = <T extends object>(initialValue: T) => {
   const { inputValue, handleInputChange, updateByNameAndValue } = useInput(initialValue);
