+++ hooks/src/lib/useExpiryDate/useExpiryDate.ts
@@ -1,5 +1,5 @@
 import { ChangeEvent, KeyboardEvent, FocusEvent, useState } from "react";
-import { useInput } from "../common";
+import useInput from "../common/useInput";
 import Validator from "../utils/validator";
 import { ERROR_MESSAGE, OPTION } from "../constants";
 import formattingMonth from "../utils/formattingMonth";
